Extracted project conversion spec to its own file.
